requirements.txt
# requirements.txt
# Smart Retail Assistant — complete dependencies

# ── Web framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# ── ML ────────────────────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# ── Azure SDK ─────────────────────────────────────────────────────
azure-storage-blob==12.19.1
azure-search-documents==11.4.0
azure-core==1.30.1
azure-keyvault-secrets==4.8.0
azure-identity==1.16.0

# ── LangChain + OpenAI (Phase 3) ─────────────────────────────────
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
openai==1.30.1

# ── Database ──────────────────────────────────────────────────────
pyodbc==5.1.0
sqlalchemy==2.0.30

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6


sentence-transformers==2.7.0
faiss-cpu==1.7.4
